+++ tests/search.rs
@@ -2,5 +2,7 @@ mod common;
 
 #[path = "search/live.rs"]
 mod search_live;
+#[path = "search/names.rs"]
+mod search_names;
 #[path = "search/offline.rs"]
 mod search_offline;